refactor(runtime): remove parallel realtime seams
