+++ packages/compiler/src/runtime/index.ts
@@ -14,10 +14,7 @@ import {
 import type { ApplicationConfiguration, NormalizedResource } from "../types";
 
 export { renderClientContract, renderCodecType } from "./client";
-export {
-	projectRealtimeWireContract,
-	type RealtimeWireContractV1,
-} from "./realtime-wire";
+export { projectRealtimeWireContract } from "./realtime-wire";
 export {
 	renderApplicationBundle,
 	renderApplicationDeclaration,
