+++ packages/runtime/src/live-query/index.ts
@@ -4,45 +4,25 @@ export {
 	matchesObservedLiveQueryPlan,
 } from "./dependency-plan";
 export type {
-	LiveQueryDependencyTokenKind,
-	LiveQueryDependencyTokenV1,
 	LiveQueryObservation,
 	ObservedLiveQueryPlanV1,
 } from "./observation";
-export { createLiveQueryInvalidation } from "./invalidation";
-export type {
-	LiveQueryInvalidation,
-	LiveQueryInvalidationResultV1,
-	LiveQueryRecomputeResultV1,
-	PreparedLiveQueryInvalidationV1,
-} from "./invalidation";
 export { reconcilePostgresChangeLedger } from "./postgres";
-export type {
-	ChangeLedgerFactV1,
-	ChangeReconciliationResultV1,
-} from "./postgres";
+export type { ChangeLedgerFactV1 } from "./postgres";
 export { createPostgresLiveQueryInvalidationEffect } from "./postgres-durable-invalidation";
-export type { PostgresLiveQueryInvalidationEffect } from "./postgres-durable-invalidation";
 export {
 	createPostgresRealtimeScopeStore,
-	type PostgresRealtimeScopeStore,
 	type PostgresRealtimeScopeLease,
 	type PostgresRealtimeWatch,
 } from "./postgres-realtime-scope";
 export {
 	createPostgresLiveQueryRetention,
-	type PostgresLiveQueryRetention,
-	type RetainedLiveQueryBinding,
 	type RetainedLiveQueryCompleteResult,
 } from "./postgres-retention";
 export { createPostgresReconciliationWake } from "./postgres-wake";
-export type {
-	PostgresReconciliationWake,
-	PostgresWakeTickSource,
-} from "./postgres-wake";
+export type { PostgresWakeTickSource } from "./postgres-wake";
 export { linkLiveQueryProgram } from "./program";
 export type {
-	LinkedContextObservationSlotV1,
 	LinkedLiveQueryProgramV1,
 	LinkedQueryWatchabilityV1,
 	LinkedStructuralQueryObservationSlotV1,
