+++ packages/runtime/src/live-query/postgres-realtime-scope.ts
@@ -19,8 +19,6 @@ import {
 } from "./postgres-realtime-scope-contract";
 
 export type {
-	PostgresRealtimeGeneration,
-	PostgresRealtimeOpenWatch,
 	PostgresRealtimeScopeLease,
 	PostgresRealtimeWatch,
 } from "./postgres-realtime-scope-contract";
